🐛 zv: Don't impl Type for dicts with non-basic keys
According to the D-Bus (and GVariant) spec, the keys of a dict type can
only be a basic type so let's not allow people to construct non-conformant
types by allowing keys to be of non-basic types.

Strictly speaking, this is very much a breaking change but there is no
value in using invalid key types anyways and code doing so, will just
get disconnected by the bus w/o any errors, a compile-time error is very
much what users would prefer.

Fixes #1637.